fix: cosine_similarity returns native float (pgvector arrays → np.float32)
When a centroid comes back from pgvector as a numpy array, cosine_similarity
returned np.float32, which then failed to serialize into the Postgres events
JSONB column during growth. Cast to float() at the source.

+++ markov_engine/vectors.py
@@ -9,7 +9,9 @@ def cosine_similarity(a: list[float], b: list[float]) -> float:
     dot = sum(x * y for x, y in zip(a, b))
     na = math.sqrt(sum(x * x for x in a)) or 1.0
     nb = math.sqrt(sum(y * y for y in b)) or 1.0
-    return dot / (na * nb)
+    # Cast to a native float: when a/b come back from pgvector as numpy arrays the
+    # result is np.float32, which isn't JSON-serializable downstream (events JSONB).
+    return float(dot / (na * nb))
 
 
 def incremental_mean(centroid: list[float], new: list[float], count: int) -> list[float]:
